% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{handwriterApp}
\alias{handwriterApp}
\title{Handwriter Application}
\usage{
handwriterApp(...)
}
\arguments{
\item{...}{Other arguments passed on to arguments \code{onStart}, \code{options},
\code{uiPattern}, and \code{enableBookmarking} of \code{shiny::shinyApp}}
}
\description{
Lauch a Shiny App for the \code{handwriter} R package.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  options(device.ask.default = FALSE)
  handwriterApp()
}

}
