% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_writer_profiles}
\alias{plot_writer_profiles}
\title{Plot Writer Profiles}
\usage{
plot_writer_profiles(rates)
}
\arguments{
\item{rates}{A data frame of cluster fill rates created with
\code{\link[handwriterRF]{get_cluster_fill_rates}}}
}
\value{
A line plot
}
\description{
Create a line plot of cluster fill rates for one or more documents, where the
cluster fill rates serve as writer profiles. Each cluster fill rates for each
document are plotted as different colored lines.
}
\examples{
plot_writer_profiles(rates)

}
