% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{Cluster Fill Rates}
\format{
A data frame
\describe{
\item{docname}{The file name of the document without the file extension.}
\item{total_graphs}{The total number of graphs in the document.}
\item{cluster1}{The proportion of graphs in cluster 1}
\item{cluster2}{The proportion of graphs in cluster 2}
\item{cluster3}{The proportion of graphs in cluster 3}
\item{cluster4}{The proportion of graphs in cluster 4}
\item{cluster5}{The proportion of graphs in cluster 5}
\item{cluster6}{The proportion of graphs in cluster 6}
\item{cluster7}{The proportion of graphs in cluster 7}
\item{cluster8}{The proportion of graphs in cluster 8}
\item{cluster9}{The proportion of graphs in cluster 9}
\item{cluster10}{The proportion of graphs in cluster 10}
\item{cluster11}{The proportion of graphs in cluster 11}
\item{cluster12}{The proportion of graphs in cluster 12}
\item{cluster13}{The proportion of graphs in cluster 13}
\item{cluster14}{The proportion of graphs in cluster 14}
\item{cluster15}{The proportion of graphs in cluster 15}
\item{cluster16}{The proportion of graphs in cluster 16}
\item{cluster17}{The proportion of graphs in cluster 17}
\item{cluster18}{The proportion of graphs in cluster 18}
\item{cluster19}{The proportion of graphs in cluster 19}
\item{cluster20}{The proportion of graphs in cluster 20}
\item{cluster21}{The proportion of graphs in cluster 21}
\item{cluster22}{The proportion of graphs in cluster 22}
\item{cluster23}{The proportion of graphs in cluster 23}
\item{cluster24}{The proportion of graphs in cluster 24}
\item{cluster25}{The proportion of graphs in cluster 25}
\item{cluster26}{The proportion of graphs in cluster 26}
\item{cluster27}{The proportion of graphs in cluster 27}
\item{cluster28}{The proportion of graphs in cluster 28}
\item{cluster29}{The proportion of graphs in cluster 29}
\item{cluster30}{The proportion of graphs in cluster 30}
\item{cluster31}{The proportion of graphs in cluster 31}
\item{cluster32}{The proportion of graphs in cluster 32}
\item{cluster33}{The proportion of graphs in cluster 33}
\item{cluster34}{The proportion of graphs in cluster 34}
\item{cluster35}{The proportion of graphs in cluster 35}
\item{cluster36}{The proportion of graphs in cluster 36}
\item{cluster37}{The proportion of graphs in cluster 37}
\item{cluster38}{The proportion of graphs in cluster 38}
\item{cluster39}{The proportion of graphs in cluster 39}
\item{cluster40}{The proportion of graphs in cluster 40}}
}
\usage{
rates
}
\description{
A data frame of cluster fill rates for two handwritten documents:
w0004_s01_pLND_r01.png and w0004_s01_pWOZ_r02.png. Both documents are from
the CSAFE Handwriting Database.
}
\details{
'handwriter' splits handwriting in the documents into component shapes called
\emph{graphs}. The graphs are sorted into 40 clusters using the cluster template
'templateK40'. The rates data frame shows the proportion of graphs from each
document assigned to each cluster. The rates estimate a writer profile for
the writer of a document.
}
\examples{
plot_writer_profiles(rates)

}
\keyword{cluster}
