% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{Clustering Template}
\format{
A named list
}
\usage{
template
}
\description{
A clustering template created with \code{handwriter::make_clustering_templates}. The
template was created from 100 training documents, each from a different writer. The
template has 40 clusters.
}
\examples{
# View the number of clusters in the template
template$K
}
\keyword{datasets}
