% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templateKforty}
\alias{templateKforty}
\title{Cluster Template with Forty Clusters}
\format{
An object of class \code{list} of length 13.
}
\usage{
templateKforty
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{cluster}
