% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{handwriterApp}
\alias{handwriterApp}
\title{Handwriter Application}
\usage{
handwriterApp(...)
}
\arguments{
\item{...}{Optional arguments passed to shiny::shinyApp}
}
\value{
No return value, called to launch 'shiny' app

A Shiny app
}
\description{
Lauch a 'shiny' application for 'handwriter'.
}
\examples{
\dontrun{
handwriterApp()
}

}
\keyword{Shiny}
