% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{handwriterApp}
\alias{handwriterApp}
\title{Handwriter Application}
\usage{
handwriterApp(...)
}
\arguments{
\item{...}{Other arguments passed on to 'onStart', 'options',
'uiPattern', or 'enableBookmarking' of 'shiny::shinyApp'}
}
\value{
No return value, called to launch 'shiny' app
}
\description{
Lauch a 'shiny' application for 'handwriter'.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  options(device.ask.default = FALSE)
  handwriterApp()
}

}
