% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{handwriterApp}
\alias{handwriterApp}
\title{Handwriter App}
\usage{
handwriterApp
}
\arguments{
\item{...}{Optional arguments passed to shiny::shinyApp}
}
\value{
A Shiny app
}
\description{
An R Shiny app for the \code{handwriter} R package.
}
\examples{
\dontrun{
handwriterApp()
}

}
\keyword{Shiny}
